Spent some time researching the kthread workflow management system and reorganized this stub controller use kthread work management.
kthread APIs are different between 4.1 and 4.13 'kthread_run, queue_kthread_work, and other APIs were renamed, going to need to experiment to find the best way to support version differences. (1: do not used the changed APIs, find substitutes that are stable; 2: use #ifdef style techniques to include both versions of the APIs in the source code .)

The kernel controller right now prints to dmesg and then exits. what it should do is: for each known probe, count the milliseconds until the timeout value is equaled. A timed-out work node is flagged as a repeat work or a one-shot work. Repeat works that tinme out are linked back onto the work list where the worker thread will time it out all over again.

Some changes in the todo.md reflect a question:

What coding on the kernel controller could we do in parallel for stage 2, if we had three developers full time?