Updates to Windows sensor_winvirtue and related.
These changes don't pertain to the desktop bridge but I found a number of problems with the Windows sensor that needed to be fixed. Foremost, the sensor stopped streaming after a couple of minutes, since the underlying handle to the filter comm port was being used synchronously from a coroutine, thus screwing up the async scheduler. I believe this is fixed in the sensor, but now ntfltmgr.py dies without explanation after a couple minutes in standalone mode.